feat(profile): updates

+++ CareNestSolution/Profile/Domain/Model/Commands/DeleteProfileCommand.cs
@@ -1,5 +1,5 @@
 ﻿namespace CareNestSolution.Profile.Domain.Model.Commands;
 
 public record DeleteProfileCommand(
-    long Id
+    int Id
     );
